chore(ctr): rename post component to CtrPostView
